refactor(plugin): rename primitives to kdco-primitives
Namespace primitives module to prevent collisions when projects
use multiple registries. Each registry prefixes shared utilities
with the registry name.

- Rename primitives/ → kdco-primitives/ in plugin folders
- Update all import paths in source and test files
- Update registry manifest component name and dependencies
- Update JSDoc @module tags
