+++ package.json
@@ -1,32 +1,5 @@
 {
-	"name": "ocx",
-	"version": "1.0.0",
-	"description": "ShadCN-style CLI for OpenCode extensions - agents, skills, plugins, commands",
-	"author": "kdcokenny",
-	"license": "MIT",
-	"type": "module",
-	"private": true,
-	"packageManager": "bun@1.3.5",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/kdcokenny/ocx.git"
-	},
-	"workspaces": [
-		"packages/*",
-		"registry/*",
-		"workers/*"
-	],
-	"scripts": {
-		"build": "turbo run build",
-		"dev": "turbo run dev",
-		"check": "turbo run check",
-		"format": "biome check --write .",
-		"test": "turbo run test"
-	},
-	"devDependencies": {
-		"@biomejs/biome": "^2.3.10",
-		"turbo": "^2.5.4",
-		"typescript": "^5.8.3",
-		"wrangler": "^4.54.0"
-	}
+  "dependencies": {
+    "@opencode-ai/plugin": "1.1.14"
+  }
 }
