+++ workers/kdco-registry/files/plugin/workspace-plugin.ts
@@ -3,7 +3,7 @@ import * as os from "node:os"
 import * as path from "node:path"
 import { type Plugin, tool } from "@opencode-ai/plugin"
 import { z } from "zod"
-import { getProjectId } from "./primitives/get-project-id"
+import { getProjectId } from "./kdco-primitives/get-project-id"
 
 // ==========================================
 // PLAN SCHEMA & VALIDATION
