+++ workers/kdco-registry/files/plugin/worktree.ts
@@ -17,7 +17,7 @@ import * as os from "node:os"
 import * as path from "node:path"
 import { type Plugin, tool } from "@opencode-ai/plugin"
 import type { Event } from "@opencode-ai/sdk"
-import type { OpencodeClient } from "./primitives/types"
+import type { OpencodeClient } from "./kdco-primitives/types"
 
 /** Logger interface for structured logging */
 interface Logger {
@@ -30,7 +30,7 @@ interface Logger {
 import { parse as parseJsonc } from "jsonc-parser"
 import { z } from "zod"
 
-import { getProjectId } from "./primitives/get-project-id"
+import { getProjectId } from "./kdco-primitives/get-project-id"
 import {
 	addSession,
 	clearPendingDelete,
