+++ workers/kdco-registry/tests/worktree-concurrent.test.ts
@@ -13,7 +13,7 @@ import { afterEach, beforeEach, describe, expect, it } from "bun:test"
 import fs from "node:fs"
 import os from "node:os"
 import path from "node:path"
-import { Mutex } from "../files/plugin/primitives/mutex"
+import { Mutex } from "../files/plugin/kdco-primitives/mutex"
 import {
 	addSession,
 	clearPendingSpawn,
