+++ workers/kdco-registry/tests/worktree-terminal.test.ts
@@ -7,7 +7,7 @@ import { afterEach, beforeEach, describe, expect, it } from "bun:test"
 import fs from "node:fs"
 import os from "node:os"
 import path from "node:path"
-import { escapeAppleScript, escapeBash, escapeBatch } from "../files/plugin/primitives/shell"
+import { escapeAppleScript, escapeBash, escapeBatch } from "../files/plugin/kdco-primitives/shell"
 import { withTempScript } from "../files/plugin/worktree/terminal"
 
 describe("worktree-terminal", () => {
